backend/requirements.txt
fastapi==0.104.1
pydantic==2.5.1
motor==3.3.2  # asynchronous MongoDB driver for Python
dnspython==2.4.2  # used to resolve the DNS of our MongoDB instance
uvicorn==0.24.0.post1
pytest==7.4.3
python-decouple==3.8
isort==5.12.0  # automatically sorting and formatting import statements 
pre-commit==3.5.0
httpie==3.2.2  # a command-line REST API client to test API endpoints in the terminal
pytest-asyncio==0.21.1